requirements.txt
kafka-python
requests
streamlit
pandas
google-cloud-bigquery
db-dtypes
google-cloud-bigquery-storage